Move five JSPs plus notification.txt to specimen-main Move NotificationBean and SpecimenRequestNotificationEmailTemplate
