+++ devtools/src/org/labkey/devtools/ActionListView.java
@@ -1,5 +1,6 @@
 package org.labkey.devtools;
 
+import org.labkey.api.action.FormHandlerAction;
 import org.labkey.api.action.SpringActionController;
 import org.labkey.api.data.Container;
 import org.labkey.api.util.PageFlowUtil;
@@ -7,9 +8,7 @@ import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.HttpView;
 
 import java.io.PrintWriter;
-import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.List;
 
 public class ActionListView extends HttpView
 {
@@ -23,14 +22,14 @@ public class ActionListView extends HttpView
     @Override
     protected void renderInternal(Object model, PrintWriter out)
     {
-        List<SpringActionController.ActionDescriptor> descriptors = new ArrayList<>(_controller.getActionResolver().getActionDescriptors());
-        descriptors.sort(Comparator.comparing(SpringActionController.ActionDescriptor::getPrimaryName));
         Container c = _controller.getViewContext().getContainer();
 
-        for (SpringActionController.ActionDescriptor ad : descriptors)
-        {
-            out.println(PageFlowUtil.link(ad.getPrimaryName()).href(new ActionURL(ad.getActionClass(), c)).clearClasses());
-            out.print("<br>");
-        }
+        _controller.getActionResolver().getActionDescriptors().stream()
+            .filter(ad->!(FormHandlerAction.class.isAssignableFrom(ad.getActionClass())))
+            .sorted(Comparator.comparing(SpringActionController.ActionDescriptor::getPrimaryName))
+            .forEach(ad->{
+                out.println(PageFlowUtil.link(ad.getPrimaryName()).href(new ActionURL(ad.getActionClass(), c)).clearClasses());
+                out.print("<br>");
+            });
     }
 }
