+++ study/src/org/labkey/study/controllers/specimen/SpecimenUtils.java
@@ -17,7 +17,6 @@
 package org.labkey.study.controllers.specimen;
 
 import org.apache.commons.lang3.StringUtils;
-import org.jetbrains.annotations.NotNull;
 import org.labkey.api.action.SpringActionController;
 import org.labkey.api.annotations.Migrate;
 import org.labkey.api.attachments.Attachment;
@@ -51,7 +50,6 @@ import org.labkey.api.specimen.RequestEventType;
 import org.labkey.api.specimen.RequestedSpecimens;
 import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.SpecimenRequestManager;
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.actions.VialRequestForm;
@@ -60,7 +58,9 @@ import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.specimen.notifications.ActorNotificationRecipientSet;
+import org.labkey.api.specimen.notifications.DefaultRequestNotification;
 import org.labkey.api.specimen.notifications.NotificationRecipientSet;
+import org.labkey.api.specimen.query.SpecimenQueryView;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
@@ -70,10 +70,12 @@ import org.labkey.api.specimen.security.permissions.SetSpecimenCommentsPermissio
 import org.labkey.api.specimen.settings.RepositorySettings;
 import org.labkey.api.specimen.settings.RequestNotificationSettings;
 import org.labkey.api.specimen.settings.SettingsManager;
+import org.labkey.api.specimen.view.NotificationBean;
+import org.labkey.api.specimen.view.SpecimenRequestNotificationEmailTemplate;
 import org.labkey.api.study.CohortFilter;
-import org.labkey.api.study.Location;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
+import org.labkey.api.study.model.ParticipantDataset;
 import org.labkey.api.util.Button;
 import org.labkey.api.util.DateUtil;
 import org.labkey.api.util.MailHelper;
@@ -88,14 +90,10 @@ import org.labkey.api.view.ViewContext;
 import org.labkey.study.controllers.BaseStudyController;
 import org.labkey.study.controllers.StudyController;
 import org.labkey.study.model.DatasetDefinition;
-import org.labkey.api.study.model.ParticipantDataset;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
-import org.labkey.api.specimen.query.SpecimenQueryView;
 import org.labkey.study.query.StudyQuerySchema;
-import org.labkey.api.specimen.notifications.DefaultRequestNotification;
-import org.labkey.study.view.specimen.SpecimenRequestNotificationEmailTemplate;
 import org.springframework.validation.BindException;
 import org.springframework.web.servlet.mvc.Controller;
 
@@ -525,110 +523,6 @@ public class SpecimenUtils
             else
                 SpecimenRequestManager.get().createRequestEvent(getUser(), specimenRequest,
                         RequestEventType.NOTIFICATION_SENT, "Notification sent to " + recipient.getLongRecipientDescription(), null);
-        }
-    }
-
-    public static class NotificationBean
-    {
-        private final DefaultRequestNotification _notification;
-        private final User _user;
-        private final String _specimenList;
-        private final String _studyName;
-        private final String _requestURI;
-
-        private boolean _includeSpecimensInBody;
-
-        public NotificationBean(ViewContext context, DefaultRequestNotification notification, String specimenList, String studyName)
-        {
-            _notification = notification;
-            _user = context.getUser();
-            _specimenList = specimenList;
-            _studyName = studyName;
-            _requestURI = new ActionURL(SpecimenController.ManageRequestAction.class, context.getContainer()).getURIString();
-
-            _includeSpecimensInBody = null != specimenList;
-        }
-
-        public @NotNull List<Attachment> getAttachments()
-        {
-            return _notification.getAttachments();
-        }
-
-        public String getComments()
-        {
-            return _notification.getComments();
-        }
-
-        public int getRequestId()
-        {
-            return _notification.getSpecimenRequest().getRowId();
-        }
-
-        public String getModifyingUser()
-        {
-            return _user.getEmail();
-        }
-
-        public String getRequestingSiteName()
-        {
-            Location destLocation = LocationManager.get().getLocation(_notification.getSpecimenRequest().getContainer(),
-                    _notification.getSpecimenRequest().getDestinationSiteId());
-            if (destLocation != null)
-                return destLocation.getDisplayName();
-            else
-                return null;
-        }
-
-        public String getStatus()
-        {
-            SpecimenRequestStatus status = SpecimenRequestManager.get().getRequestStatus(_notification.getSpecimenRequest().getContainer(),
-                    _notification.getSpecimenRequest().getStatusId());
-            return status != null ? status.getLabel() : "Unknown";
-        }
-
-        public Container getContainer()
-        {
-            return _notification.getSpecimenRequest().getContainer();
-        }
-
-        public String getEventDescription()
-        {
-            return _notification.getEventSummary();
-        }
-
-        public String getRequestDescription()
-        {
-            return _notification.getSpecimenRequest().getComments();
-        }
-
-        public String getSpecimenList()
-        {
-            return _specimenList;
-        }
-
-        public String getStudyName()
-        {
-            return _studyName;
-        }
-
-        public String getRequestURI()
-        {
-            return _requestURI;
-        }
-
-        public boolean getIncludeSpecimensInBody()
-        {
-            return _includeSpecimensInBody;
-        }
-
-        public void setIncludeSpecimensInBody(boolean includeSpecimensInBody)
-        {
-            _includeSpecimensInBody = includeSpecimensInBody;
-        }
-
-        public SpecimenRequestEvent getEvent()
-        {
-            return _notification.getEvent();
         }
     }
