+++ study/src/org/labkey/study/view/specimen/manageActorOrder.jsp
@@ -17,15 +17,11 @@
 %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenRequestActor"%>
 <%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider"%>
-<%@ page import="org.labkey.api.view.HttpView"%>
-<%@ page import="org.labkey.api.view.JspView"%>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageActorOrderAction"%>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageActorsAction" %>
-<%@ page import="org.labkey.study.model.StudyImpl" %>
+<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageActorsAction"%>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
-    JspView<StudyImpl> me = (JspView<StudyImpl>) HttpView.currentView();
     SpecimenRequestActor[] actors = SpecimenRequestRequirementProvider.get().getActors(getContainer());
 %>
 <labkey:errors/>
