+++ study/src/org/labkey/study/view/specimen/manageRepositorySettings.jsp
@@ -16,11 +16,11 @@
  */
 %>
 <%@ page import="org.labkey.api.query.QueryUrls" %>
+<%@ page import="org.labkey.api.specimen.settings.RepositorySettings" %>
 <%@ page import="org.labkey.api.study.Study" %>
 <%@ page import="org.labkey.api.study.StudyService" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageRepositorySettingsAction" %>
-<%@ page import="org.labkey.api.specimen.settings.RepositorySettings" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.study.view.BaseStudyPage" %>
 <%
