+++ study/src/org/labkey/study/view/specimen/viewRequestsHeader.jsp
@@ -15,32 +15,32 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.specimen.SpecimenRequestStatus" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.ViewContext" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController" %>
-<%@ page import="org.labkey.api.specimen.SpecimenRequestStatus" %>
+<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ViewRequestsHeaderBean" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
-    JspView<SpecimenController.ViewRequestsHeaderBean> me = (JspView<SpecimenController.ViewRequestsHeaderBean>) HttpView.currentView();
+    JspView<ViewRequestsHeaderBean> me = (JspView<ViewRequestsHeaderBean>) HttpView.currentView();
     ViewContext context = getViewContext();
-    SpecimenController.ViewRequestsHeaderBean bean = me.getModelBean();
+    ViewRequestsHeaderBean bean = me.getModelBean();
     ActionURL userLink = context.cloneActionURL();
 %>
-<%= link("All User Requests", userLink.deleteParameter(SpecimenController.ViewRequestsHeaderBean.PARAM_CREATEDBY)) %>
-<%= link("My Requests", userLink.replaceParameter(SpecimenController.ViewRequestsHeaderBean.PARAM_CREATEDBY,
+<%= link("All User Requests", userLink.deleteParameter(ViewRequestsHeaderBean.PARAM_CREATEDBY)) %>
+<%= link("My Requests", userLink.replaceParameter(ViewRequestsHeaderBean.PARAM_CREATEDBY,
         getUser().getDisplayName(getUser()))) %>
 Filter by status: <select onChange="document.location=options[selectedIndex].value;">
 <%
     ActionURL current = context.cloneActionURL();
-    current.deleteParameter(SpecimenController.ViewRequestsHeaderBean.PARAM_STATUSLABEL);
+    current.deleteParameter(ViewRequestsHeaderBean.PARAM_STATUSLABEL);
 %>
     <option value="<%= h(current.getLocalURIString()) %>">All Statuses</option>
 <%
     for (SpecimenRequestStatus status : bean.getStauses())
     {
-        current.replaceParameter(SpecimenController.ViewRequestsHeaderBean.PARAM_STATUSLABEL, status.getLabel());
+        current.replaceParameter(ViewRequestsHeaderBean.PARAM_STATUSLABEL, status.getLabel());
 %>
     <option value="<%= h(current.getLocalURIString()) %>" <%=selected(bean.isFilteredStatus(status))%>><%= h(status.getLabel()) %></option>
 <%
@@ -48,7 +48,7 @@ Filter by status: <select onChange="document.location=options[selectedIndex].val
 %>
 </select>
 <%
-    String userFilter = getActionURL().getParameter(SpecimenController.ViewRequestsHeaderBean.PARAM_CREATEDBY);
+    String userFilter = getActionURL().getParameter(ViewRequestsHeaderBean.PARAM_CREATEDBY);
     if (userFilter != null)
     {
 %>
